Move ByteStr compare and index traits to a separate module
This parallels the layout of `core::str`.
